# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mkutano Convening Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
        - name: workstreamId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
        - name: conveningId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ── Analytics (dashboard wow endpoint) ──────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Portal User / RBAC ──────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, conveningId, workstreamId, title, status, progressPct, isMilestone]
      properties:
        id:
          type: string
        conveningId:
          type: string
        workstreamId:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        assignee:
          type: [string, 'null']
        dueDate:
          type: [string, 'null']
          format: date
        startDate:
          type: [string, 'null']
          format: date
        progressPct:
          type: integer
          minimum: 0
          maximum: 100
        isMilestone:
          type: boolean
        milestoneId:
          type: [string, 'null']
        milestoneName:
          type: [string, 'null']
        ownerUserId:
          type: [string, 'null']
        assigneeUserId:
          type: [string, 'null']
        assigneeName:
          type: [string, 'null']
        status:
          type: string
          enum: [NotStarted, InProgress, Blocked, Completed]
        priority:
          type: string
          enum: [Low, Medium, High, Urgent]
    TaskInput:
      type: object
      required: [conveningId, workstreamId, title]
      properties:
        conveningId:
          type: string
        workstreamId:
          type: string
        title:
          type: string
        description:
          type: string
        assignee:
          type: string
        dueDate:
          type: string
          format: date
        startDate:
          type: string
          format: date
        progressPct:
          type: integer
          minimum: 0
          maximum: 100
        isMilestone:
          type: boolean
        milestoneId:
          type: string
        ownerUserId:
          type: string
        assigneeUserId:
          type: string
        status:
          type: string
          enum: [NotStarted, InProgress, Blocked, Completed]
        priority:
          type: string
          enum: [Low, Medium, High, Urgent]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: [string, 'null']
        workstreamId:
          type: string
        assignee:
          type: [string, 'null']
        dueDate:
          type: [string, 'null']
          format: date
        startDate:
          type: [string, 'null']
          format: date
        progressPct:
          type: integer
          minimum: 0
          maximum: 100
        isMilestone:
          type: boolean
        milestoneId:
          type: [string, 'null']
        ownerUserId:
          type: [string, 'null']
        assigneeUserId:
          type: [string, 'null']
        status:
          type: string
          enum: [NotStarted, InProgress, Blocked, Completed]
        priority:
          type: string
          enum: [Low, Medium, High, Urgent]

    # ── Task Attachments ─────────────────────────────────────────────────────
...
    UploadUrlRequest:
      type: object
      required: [conveningId, name, size, contentType]
      properties:
        conveningId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name:
              type: string
            size:
              type: number
            contentType:
              type: string

    # ── Delegates ──────────────────────────────────────────────────────────
...